requirements.txt
astroid==2.6.2
autopep8==1.5.7
CacheControl==0.12.6
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
deprecation==2.1.0
Flask==2.0.1
idna==2.10
isort==5.9.2
itsdangerous==2.0.1
Jinja2==3.0.1
jsonpickle==0.9.6
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
msgpack==1.0.2
packaging==21.0
Pillow==8.3.1
pycodestyle==2.7.0
pylint==2.9.3
pyparsing==2.4.7
python-dateutil==2.8.1
requests==2.25.1
six==1.16.0
squareup==12.0.0.20210616
toml==0.10.2
treepoem==3.10.0
urllib3==1.26.6
Werkzeug==2.0.1
wrapt==1.12.1
